fixed some warnings
svn path=/trunk/; revision=18820

+++ reactos/subsys/system/sndvol32/mixer.c
@@ -369,7 +369,7 @@ SndMixerGetProductName(PSND_MIXER Mixer,
 {
     if (Mixer->hmx)
     {
-        int lnsz = lstrlen(Mixer->Caps.szPname);
+        UINT lnsz = (UINT) lstrlen(Mixer->Caps.szPname);
         if(lnsz + 1 > uSize)
         {
             return lnsz + 1;
@@ -409,7 +409,7 @@ SndMixerGetLineName(PSND_MIXER Mixer,
 
         if (lpl != NULL)
         {
-            lnsz = lstrlen(LongName ? lpl->szName : lpl->szShortName);
+            lnsz = (UINT) lstrlen(LongName ? lpl->szName : lpl->szShortName);
             if(lnsz + 1 > uSize)
             {
                 return lnsz + 1;
